.github/assign-reviewers/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile dev-requirements.in
#
appdirs==1.4.4
    # via black
async-generator==1.10
    # via
    #   -r requirements.txt
    #   trio
attrs==20.3.0
    # via
    #   -r requirements.txt
    #   outcome
    #   trio
black==20.8b1
    # via -r dev-requirements.in
certifi==2020.11.8
    # via
    #   -r requirements.txt
    #   httpx
cffi==1.14.4
    # via
    #   -r requirements.txt
    #   cryptography
click==7.1.2
    # via
    #   black
    #   pip-tools
cryptography==3.3.2
    # via
    #   -r requirements.txt
    #   pyjwt
gidgethub[httpx]==4.2.0
    # via -r requirements.txt
h11==0.11.0
    # via
    #   -r requirements.txt
    #   httpcore
httpcore==0.12.2
    # via
    #   -r requirements.txt
    #   httpx
httpx==0.16.1
    # via
    #   -r requirements.txt
    #   gidgethub
idna==2.10
    # via
    #   -r requirements.txt
    #   rfc3986
    #   trio
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
    #   trio-typing
mypy==0.790
    # via trio-typing
outcome==1.1.0
    # via
    #   -r requirements.txt
    #   trio
pathspec==0.8.1
    # via black
pip-tools==5.4.0
    # via -r dev-requirements.in
pycparser==2.20
    # via
    #   -r requirements.txt
    #   cffi
pyjwt[crypto]==1.7.1
    # via
    #   -r requirements.txt
    #   gidgethub
pyyaml==5.3.1
    # via -r requirements.txt
regex==2020.11.13
    # via black
rfc3986[idna2008]==1.4.0
    # via
    #   -r requirements.txt
    #   httpx
six==1.15.0
    # via
    #   -r requirements.txt
    #   cryptography
    #   pip-tools
sniffio==1.2.0
    # via
    #   -r requirements.txt
    #   httpcore
    #   httpx
    #   trio
sortedcontainers==2.3.0
    # via
    #   -r requirements.txt
    #   trio
toml==0.10.2
    # via black
trio-typing==0.5.0
    # via -r dev-requirements.in
trio==0.17.0
    # via
    #   -r requirements.txt
    #   trio-typing
typed-ast==1.4.1
    # via
    #   black
    #   mypy
typing-extensions==3.7.4.3
    # via
    #   black
    #   mypy
    #   trio-typing
uritemplate==3.0.1
    # via
    #   -r requirements.txt
    #   gidgethub
wheel==0.35.1
    # via -r requirements.txt

# The following packages are considered to be unsafe in a requirements file:
# pip
